Tried Updating The score for Level2
I have an idea but don't know how to write it in code. I wrote it, but it doesn't work.

Basically I want to measure the distance between the two notes:
The static note and non static note.

For 2nd intervals - Major Second and Minor Second.
 if  the distance between them is  " < 3 " then add one point . And then update the score on the screen after the notes are destroyed.

For Major and Minors 3rds - the distance would have to be greater than -3 but less then - 6  etc.

I ALMOST FINISHED THE GAME though.